Studio Works 3D Drawings updated, initial commit
Started working on the PrISM64 case in SW.  I want to keep the materials
hacker/maker friendly, so am trying to use sheet metal, acrylic, wood,
etc.  The top acrylic is cut from laser cutter, and laid onto a form and
heated with a heat gun to get the nice bends.